requirements.txt
python-telegram-bot==21.0.1
aiosqlite==0.20.0
pytz==2024.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
cryptography==42.0.5
PyPDF2==3.0.1
